refactor(step2): move files
