+++ src/main/kotlin/calculator/expression/StringExpression.kt
@@ -1,4 +1,4 @@
-package calculator
+package calculator.expression
 
 class StringExpression(
     expression: String?
